refactor: move test helpers
