+++ test/package/environments/base.test.js
@@ -3,7 +3,7 @@
 // environment.js expects to find config/shakapacker.yml and resolved modules from
 // the root of a Rails project
 
-const { chdirTestApp, chdirCwd, resetEnv } = require('../../../package/utils/helpers')
+const { chdirTestApp, chdirCwd, resetEnv } = require('../../helpers')
 
 const rootPath = process.cwd()
 chdirTestApp()
