+++ test/package/environments/production.test.js
@@ -1,6 +1,6 @@
 /* global test expect, describe, afterAll, beforeEach */
 
-const { chdirTestApp, resetEnv } = require('../../../package/utils/helpers')
+const { chdirTestApp, resetEnv } = require('../../helpers')
 const rootPath = process.cwd()
 chdirTestApp()
