+++ test/package/index.test.js
@@ -1,5 +1,5 @@
 const index = require('../../package/index')
-const { generateWebpackConfig } = require("../../package/index");
+const { generateWebpackConfig } = require('../helpers')
 
 describe('index', () => {
   test('exports webpack-merge v5 functions', () => {
